Grow Wild: An Interactive, Proximity Sensitive Sign
A friend of mine had one of his moss graffiti signs exhibited and asked
me to pimp it up a bit. We used an RGB LED strip connected to an
Arduino and an ultrasonic distance sensor to make the sign glow more or
less depending on how close a spectator comes. Find more detailed
description and source code of the project here:
http://kavasmlikon.wordpress.com/2012/03/24/grow-wild-an-interactive-pro
ximity-sensitive-sign/